Pull Request for SDG Icons (#2)
* Initial commit.

* refactor: Convert function names from camel case
to snake case

* fix: implement dcc.Store to replace global
variables

BREAKING CHANGE: Global variables are no longer used to store data in
the Dash app, except for the mongo uri

* refactor: remove print statements, exception handling

* style: remove comments related to user authentication

* feat: connect to MongoDB without storing URI as a global variable

* refactor: separated callbacks from app.py

* refactor: renamed script.py to database.py

* refactor: moved paragraph selection logic out of callbacks

* fix: sorting and limit in get_paragraph function

* feat: implemented paragraph time control

Now paragraphs being shown to the user are put in a queue in a separate collection.
It now takes an hour minimum for a paragraph to be shown again.

* refactor: Convert global variables to functions for layout components

* refactor: Convert callback function names to snake case

* fix: refine requirements.txt file

* chore: add SDG icons to assets folder

* feat: SDG chips now have icons

NOTE: This commit does not bring fuctionality to the icons, it only
adds them to the chips. The icons are not yet clickable.

* feat: new SDG icons that work as the precious sdg chips

NOTE: This commit has edited out some communication to the database for the sake
of debugging, this will be changed in the future

* perf: recent document ids are now stored in dcc.Store

This improves performance by considerably reducing database queries

* refactor: merged dev branch into feature-sdg-icons

* feat: implemented user email input

* feat: minor changes to text being shown in the app

* feat: changed cursor type for sdg icons

* feat: quit button now seds user to the finish layout

* feat: link to SDG info page

* fix: support for null label arrays in pipeline selection

Previously, the pipeline would filter out documents with null label arrays, but this behavior was not intended. The bug has been fixed by adding support for null label arrays, which are now treated as having a label count of zero. This change allows all documents to be selected by the pipeline.

* fix: Prevent users from labelling the same paragraph more than once

Previously, there was a bug that allowed users to label the same paragraph multiple times. This has been fixed by implementing a check for existing labels before allowing a new label to be added, using the user's email

* fix: prevent users from labeling the same paragraph in different sessions

User email control is being used to ensure the same user is not shown the same paragraph in different sessions

* fix: less restrictive requirements

* refactor: SDG icons now use snake case

* Fix: add dash-iconify to requirements.txt

* Fix: use env variables for database connection

---------

Co-authored-by: Stiene Praet <stiene.praet@undp.org>